requirements.txt
absl-py==1.0.0
attrs==21.4.0
cycler==0.11.0
fonttools==4.29.1
keyboard==0.13.5
kiwisolver==1.3.2
matplotlib==3.5.1
mediapipe==0.8.9.1
numpy==1.22.1
opencv-contrib-python==4.5.5.62
opencv-python==4.5.5.62
packaging==21.3
Pillow==9.0.0
protobuf==3.19.4
pyobjc==8.2
pyobjc-core==8.2
pyobjc-framework-Accessibility==8.2
pyobjc-framework-Accounts==8.2
pyobjc-framework-AddressBook==8.2
pyobjc-framework-AdServices==8.2
pyobjc-framework-AdSupport==8.2
pyobjc-framework-AppleScriptKit==8.2
pyobjc-framework-AppleScriptObjC==8.2
pyobjc-framework-ApplicationServices==8.2
pyobjc-framework-AppTrackingTransparency==8.2
pyobjc-framework-AudioVideoBridging==8.2
pyobjc-framework-AuthenticationServices==8.2
pyobjc-framework-AutomaticAssessmentConfiguration==8.2
pyobjc-framework-Automator==8.2
pyobjc-framework-AVFoundation==8.2
pyobjc-framework-AVKit==8.2
pyobjc-framework-BusinessChat==8.2
pyobjc-framework-CalendarStore==8.2
pyobjc-framework-CallKit==8.2
pyobjc-framework-CFNetwork==8.2
pyobjc-framework-ClassKit==8.2
pyobjc-framework-CloudKit==8.2
pyobjc-framework-Cocoa==8.2
pyobjc-framework-Collaboration==8.2
pyobjc-framework-ColorSync==8.2
pyobjc-framework-Contacts==8.2
pyobjc-framework-ContactsUI==8.2
pyobjc-framework-CoreAudio==8.2
pyobjc-framework-CoreAudioKit==8.2
pyobjc-framework-CoreBluetooth==8.2
pyobjc-framework-CoreData==8.2
pyobjc-framework-CoreHaptics==8.2
pyobjc-framework-CoreLocation==8.2
pyobjc-framework-CoreMedia==8.2
pyobjc-framework-CoreMediaIO==8.2
pyobjc-framework-CoreMIDI==8.2
pyobjc-framework-CoreML==8.2
pyobjc-framework-CoreMotion==8.2
pyobjc-framework-CoreServices==8.2
pyobjc-framework-CoreSpotlight==8.2
pyobjc-framework-CoreText==8.2
pyobjc-framework-CoreWLAN==8.2
pyobjc-framework-CryptoTokenKit==8.2
pyobjc-framework-DataDetection==8.2
pyobjc-framework-DeviceCheck==8.2
pyobjc-framework-DictionaryServices==8.2
pyobjc-framework-DiscRecording==8.2
pyobjc-framework-DiscRecordingUI==8.2
pyobjc-framework-DiskArbitration==8.2
pyobjc-framework-DVDPlayback==8.2
pyobjc-framework-EventKit==8.2
pyobjc-framework-ExceptionHandling==8.2
pyobjc-framework-ExecutionPolicy==8.2
pyobjc-framework-ExternalAccessory==8.2
pyobjc-framework-FileProvider==8.2
pyobjc-framework-FileProviderUI==8.2
pyobjc-framework-FinderSync==8.2
pyobjc-framework-FSEvents==8.2
pyobjc-framework-GameCenter==8.2
pyobjc-framework-GameController==8.2
pyobjc-framework-GameKit==8.2
pyobjc-framework-GameplayKit==8.2
pyobjc-framework-ImageCaptureCore==8.2
pyobjc-framework-IMServicePlugIn==8.2
pyobjc-framework-InputMethodKit==8.2
pyobjc-framework-InstallerPlugins==8.2
pyobjc-framework-InstantMessage==8.2
pyobjc-framework-Intents==8.2
pyobjc-framework-IntentsUI==8.2
pyobjc-framework-IOSurface==8.2
pyobjc-framework-iTunesLibrary==8.2
pyobjc-framework-KernelManagement==8.2
pyobjc-framework-LatentSemanticMapping==8.2
pyobjc-framework-LaunchServices==8.2
pyobjc-framework-libdispatch==8.2
pyobjc-framework-LinkPresentation==8.2
pyobjc-framework-LocalAuthentication==8.2
pyobjc-framework-LocalAuthenticationEmbeddedUI==8.2
pyobjc-framework-MailKit==8.2
pyobjc-framework-MapKit==8.2
pyobjc-framework-MediaAccessibility==8.2
pyobjc-framework-MediaLibrary==8.2
pyobjc-framework-MediaPlayer==8.2
pyobjc-framework-MediaToolbox==8.2
pyobjc-framework-Metal==8.2
pyobjc-framework-MetalKit==8.2
pyobjc-framework-MetalPerformanceShaders==8.2
pyobjc-framework-MetalPerformanceShadersGraph==8.2
pyobjc-framework-MetricKit==8.2
pyobjc-framework-MLCompute==8.2
pyobjc-framework-ModelIO==8.2
pyobjc-framework-MultipeerConnectivity==8.2
pyobjc-framework-NaturalLanguage==8.2
pyobjc-framework-NetFS==8.2
pyobjc-framework-Network==8.2
pyobjc-framework-NetworkExtension==8.2
pyobjc-framework-NotificationCenter==8.2
pyobjc-framework-OpenDirectory==8.2
pyobjc-framework-OSAKit==8.2
pyobjc-framework-OSLog==8.2
pyobjc-framework-PassKit==8.2
pyobjc-framework-PencilKit==8.2
pyobjc-framework-Photos==8.2
pyobjc-framework-PhotosUI==8.2
pyobjc-framework-PreferencePanes==8.2
pyobjc-framework-PushKit==8.2
pyobjc-framework-Quartz==8.2
pyobjc-framework-QuickLookThumbnailing==8.2
pyobjc-framework-ReplayKit==8.2
pyobjc-framework-SafariServices==8.2
pyobjc-framework-SceneKit==8.2
pyobjc-framework-ScreenSaver==8.2
pyobjc-framework-ScreenTime==8.2
pyobjc-framework-ScriptingBridge==8.2
pyobjc-framework-SearchKit==8.2
pyobjc-framework-Security==8.2
pyobjc-framework-SecurityFoundation==8.2
pyobjc-framework-SecurityInterface==8.2
pyobjc-framework-ServiceManagement==8.2
pyobjc-framework-ShazamKit==8.2
pyobjc-framework-Social==8.2
pyobjc-framework-SoundAnalysis==8.2
pyobjc-framework-Speech==8.2
pyobjc-framework-SpriteKit==8.2
pyobjc-framework-StoreKit==8.2
pyobjc-framework-SyncServices==8.2
pyobjc-framework-SystemConfiguration==8.2
pyobjc-framework-SystemExtensions==8.2
pyobjc-framework-UniformTypeIdentifiers==8.2
pyobjc-framework-UserNotifications==8.2
pyobjc-framework-UserNotificationsUI==8.2
pyobjc-framework-VideoSubscriberAccount==8.2
pyobjc-framework-VideoToolbox==8.2
pyobjc-framework-Virtualization==8.2
pyobjc-framework-Vision==8.2
pyobjc-framework-WebKit==8.2
pyparsing==3.0.7
pyserial==3.5
python-dateutil==2.8.2
six==1.16.0
